linked_list_get now handles absence of key
Prior to this modification, an error occurred when trying to retrieve a
value by a non-existent key.

Following the wise men's piece of advice, the hash table's buckets are
now NULL-initialized as well.

Also, the hash function now returns an unigned int, since it's used as
an array index anyway, no point in making it unsigned.